+++ src/timeline/Permissions.cpp
@@ -66,24 +66,24 @@ Permissions::canSend(int eventType)
 int
 Permissions::defaultLevel()
 {
-    return pl.users_default;
+    return static_cast<int>(pl.users_default);
 }
 int
 Permissions::redactLevel()
 {
-    return pl.redact;
+    return static_cast<int>(pl.redact);
 }
 int
 Permissions::changeLevel(int eventType)
 {
-    return pl.state_level(to_string(
-      qml_mtx_events::fromRoomEventType(static_cast<qml_mtx_events::EventType>(eventType))));
+    return static_cast<int>(pl.state_level(to_string(
+      qml_mtx_events::fromRoomEventType(static_cast<qml_mtx_events::EventType>(eventType)))));
 }
 int
 Permissions::sendLevel(int eventType)
 {
-    return pl.event_level(to_string(
-      qml_mtx_events::fromRoomEventType(static_cast<qml_mtx_events::EventType>(eventType))));
+    return static_cast<int>(pl.event_level(to_string(
+      qml_mtx_events::fromRoomEventType(static_cast<qml_mtx_events::EventType>(eventType)))));
 }
 
 bool
