+++ src/ui/MxcMediaProxy.cpp
@@ -122,7 +122,7 @@ MxcMediaProxy::startDownload()
             QByteArray ba = device.readAll();
             std::string temp(ba.constData(), ba.size());
             temp = mtx::crypto::to_string(mtx::crypto::decrypt_file(temp, encryptionInfo.value()));
-            buffer.setData(temp.data(), temp.size());
+            buffer.setData(temp.data(), static_cast<int>(temp.size()));
         } else {
             buffer.setData(device.readAll());
         }
